requirements.txt
# PyTorch - Install separately with:
# pip install torch==2.8.0 torchvision==0.23.0 --index-url https://download.pytorch.org/whl/cu129

# Core packages
numpy>=2.0.0
pandas>=2.0.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
scikit-learn>=1.5.0
albumentations>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Training
tensorboard>=2.14.0
tqdm>=4.66.0

# Data
kaggle>=1.5.0

# Development
jupyter>=1.0.0
ipykernel>=6.25.0

# Utils
PyYAML>=6.0.0
python-dotenv>=1.0.0

# For Vision Transformer
timm>=0.9.0

# For Vision-Language Model (CLIP)
transformers>=4.30.0
ftfy>=6.1.0

PyYAML>=6.0 